app/backend/requirements.txt
# Runtime deps for the FastAPI app (api.py) and its import chain.
# Omitted: tqdm (setup/ingest scripts only),
# requests (not imported directly; pulled in by kaggle if needed).
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
python-multipart>=0.0.9
google-genai>=1.0.0
chromadb>=0.4.0
numpy>=1.24.0
pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-socketio[asyncio-client]>=5.11.0
kaggle>=1.6.0
